[Docs] Switch to MkDocs
Part of the motivation here is just so the docs can use "admonitions",
or text boxes with warning / notice labels at the top. MkDocs can do
that with some small configuration changes. mdBook requires custom code.

It would probably be better to use something richer like AsciiDoc for
the markup, but this works, and I couldn't get an AsciiDoc system to
work.

GitOrigin-RevId: 7a8b2f7380f3a81eecfe66494d990941fd9361f6