Write VADPCM decoder
This is just an initial implementation of a VADPCM decoder. The core
decoder itself is written in pure C to make it easier to link into other
projects if desired. The decoder can successfully decode VADPCM to AIFF.
The code may be a bit rough.

The numSampleFrames field in the COMM chunk is probably not handled
correctly. It is basically ignored, and copied from input to output. It
should be correctly parsed and emitted.

The decoder was tested manually with files not checked-in. Automated
tests should be added.

GitOrigin-RevId: 3784e5e7ee772ef44a483a045a0e1e5987aae099