Allocate memory inside encoder
The original idea behind passing in scratch memory as a parameter was
that it would be easier to package the encoder so it can run in
constrained environments like WASM. However, this is a niche use case,
and it makes the API more cumbersome. We can add compile-time options
later to change the allocator, or allow the allocator functions to be
passed in through the parameters structure.